Refactor module names from controller to DAO
